+++ Topic8/src/main/java/com/aguirre/topic8/services/ScheduleTimeService.java
@@ -1,25 +1,14 @@
 package com.aguirre.topic8.services;
 
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import com.aguirre.topic8.models.ScheduleTime;
-import com.aguirre.topic8.repositories.ScheduleTimeDao;
 
 @Service
-public class ScheduleTimeService {
+public interface ScheduleTimeService {
 
-	@Autowired
-	private ScheduleTimeDao scheduleTimeDao;
+	public void addScheduleTime(ScheduleTime scheduleTime);
 
-	public void addScheduleTime(ScheduleTime scheduleTime) {
-		scheduleTimeDao.save(scheduleTime);
-	}
+	public void deleteScheduleTime(ScheduleTime scheduleTime);
 
-	public void deleteScheduleTime(ScheduleTime scheduleTime) {
-		scheduleTimeDao.delete(scheduleTime);
-	}
-
-	public ScheduleTime getScheduleTime(Long idScheduleTime) {
-		return scheduleTimeDao.findOne(idScheduleTime);
-	}
+	public ScheduleTime getScheduleTime(Long idScheduleTime);
 }
